other_resources/slack_chatbot/requirements.txt
slack_sdk
google-cloud-storage
google-cloud-tpu
google-api-python-client
google_auth_oauthlib